requirements.txt
# Core dependencies for the project
python-telegram-bot
numpy
Pillow
gunicorn

# Dependencies for the new ML API
fastapi
uvicorn[standard]
onnxruntime
scikit-image
pydantic
python-multipart
